BugFix: use '_private' as JSON field name for 'private' fields
GitHub API uses 'private' as JSON field name, which is by chance
a Java keyword and then is not allowed as Bean field name.
In the Kohsuke's API GitHub Beans the '_private' field name is used
so it needs to be mapped to the JSON fields named 'private'.

Because of this JSON field name mismatch, all private GitHub repos
were listed as public by the Kohsuke's GitHub API.