Add tests for GHContentUpdatedResponse
The `commit` field in `GHContentUpdatedResponse` actually matches up to the `GHCommit.ShortCommitInfo` rather than `GHCommit`.  The means that most of the fields are misaligned and set to null when returned.  However, `GHCommit` can recover by populating and then returning info as normal.

This is not a great solution, but it show that this part of the API is only less than optimal,
not completely broken.