Using TCK Tested JDK builds of OpenJDK
The AdoptOpenJDK has been discontinued since July 2021. When using Zulu 
you get all the latest updated (TCK Tested) builds for all versions of OpenJDK. 

As good practice I added major fixed versions of JDK 8.0.192, 11.0.3. 

An example is when 8.0.192 passes(green) and 8 (latest) fails (red) you know it was the build version not your code that caused the build to fail.

While Temurin is the successor to AdoptOpenJDK it does not support archived versions prior to sept. 2021. Azul has all archived versions and the latest releases.