feat: Add head_commit to GHEventPayload.Push
When that class represents a commit being pushed to a branch, the list
of commits is available and useful to know what was pushed.

When it represents a tag being pushed, the list of commits can be empty.
The `head_commit` comes in handy to know more about the commit.

This change exposes `head_commit` using the getter `getHeadCommit()`.

The test data already has the data for this to work. The test case has
been updated to test for this method.

Refs: #1585