Add MAINTAIN and TRIAGE permissions to permission type
Fixes https://github.com/hub4j/github-api/issues/1671

https://community.jenkins.io/t/multibranch-pipline-fails-because-triage-enum-doesnt-exist/7800
reports that the TRIAGE permission is unknown to the GitHub api library.

https://github.com/jenkins-infra/helpdesk/issues/3617 reports that the
MAINTAINER permission is unknown to the GitHub api library.

This adds both the triage and the maintain permission to the
enumeration so that new releases of the plugins depending on
this library can be done to avoid the stack trace reported in
https://github.com/jenkins-infra/helpdesk/issues/3617

That stack trace includes:

java.lang.IllegalArgumentException: No enum constant org.kohsuke.github.GHPermissionType.MAINTAIN

Special thanks to @PierreBeitz for the test case with wiremock.