Re-Prepare request for every retry
Currently, a retried request fails if the authentication token expires between retries. This can be a common experience, as GitHub API can throttle a user for up to an hour, which is longer than the lifetime of an authentication token retrieved via the GitHub Application Installation.

This commit ensures that a recent token is used in each request by re-creating the connector request for every retry.