Restructured JwtBuilderUtil based on testing
When I went to run tests on the reflection code I ran into a number of
errors. A few were straighfoward fixes but not all.

The issue requiring the biggest change as that some JWT suite classes
were being loaded as the JwtBuilderUtil class was being loaded,
or at least before the buildJwt() method started executing.

I had to move the JWT calls into a nested class to delay the loading
of JWT suite classes until inside a try-catch where I could handle it
as expected.