Replace /issues/ with /pulls/ in sourced URL for Pull Requests.
Previously, when a pull request had been fetched, it lacked an
`owner`, and thus was following a specific codepath inside
`GHPullRequest.GetApiRoute` that was returning a URL that corresponded
to an `issue`. When that URL was then appended on for further queries,
say, with `/commits`, then Github would return a 404.

This fixes the issue by manipulating the URL manually and replacing
the wrong reference to `/issues/` with `/pulls/`.