Support team, team_add, member and membership payloads
I removed/updated some outdated files that were apparently added in the
hope of someone implementing the payloads but I preferred to start fresh
with new updated payloads.

I did a few adjustments to some existing enums to make sure we don't
have an error when a payload arrives with an unknown value. I used the
same pattern we decided to implement some time ago.